rename module exportHTML.ts to toHTML.ts
