+++ src/module-init-browser.ts
@@ -1,8 +1,8 @@
-/* import * as exportHTML from './common/export/html/exportHTML';
+/* import * as toHTML from './common/export/html/toHTML';
 // Add further Modul-Inits here, ...
 
 export function initAllModules() {
-    // exportHTML.init();
+    // toHTML.init();
     // Call further Modul-Inits here ..
 }
 */
