+++ src/plugins/import/reqif/import-reqif.vue
@@ -63,7 +63,7 @@
     import { Options, Vue } from 'vue-class-component';
     import { ReqifImporter } from '@/common/import/reqif/import-reqif';
     import { TPigItem, APackage } from '@/common/schema/pig/ts/pig-metaclasses';
-    import { stringHTML, toHTML } from '@/common/export/html/exportHTML';
+    import { stringHTML, toHTML } from '@/common/export/html/toHTML';
     import { useHtmlStore } from '@/stores/cacheStore';
     import { LOG } from '@/common/lib/helpers';
     import { Msg, IRsp } from '@/common/lib/messages';
